Clean up CCLF Update Logic
