+++ maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/service/ValidationService.java
@@ -3,30 +3,21 @@ package uk.gov.justice.laa.crime.orchestration.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import uk.gov.justice.laa.crime.commons.exception.APIClientException;
 import uk.gov.justice.laa.crime.enums.RepOrderStatus;
 import uk.gov.justice.laa.crime.exception.ValidationException;
 import uk.gov.justice.laa.crime.orchestration.dto.WorkflowRequest;
-import uk.gov.justice.laa.crime.orchestration.dto.maat.ApplicationDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.maat.RepStatusDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.maat_api.RepOrderDTO;
-import uk.gov.justice.laa.crime.orchestration.dto.maat_api.SendToCCLFDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.validation.UserSummaryDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.validation.UserValidationDTO;
 import uk.gov.justice.laa.crime.orchestration.exception.CrimeValidationException;
 import uk.gov.justice.laa.crime.orchestration.service.api.MaatCourtDataApiService;
 import uk.gov.justice.laa.crime.util.DateUtil;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
-import java.util.Objects;
 
 @Slf4j
 @Service
@@ -48,43 +39,6 @@ public class ValidationService {
             "User does not have a valid New Work Reason Code";
     private final MaatCourtDataService maatCourtDataService;
     private final MaatCourtDataApiService maatCourtDataApiService;
-
-    public static String getRepOrderCcOutcome(RepOrderDTO repOrderDTO) {
-        if (repOrderDTO.getRepOrderCCOutcome() == null || repOrderDTO.getRepOrderCCOutcome().isEmpty()) return null;
-        return repOrderDTO.getRepOrderCCOutcome().get(0).getOutcome();
-    }
-
-    public static String getAppealType(ApplicationDTO applicationDTO) {
-        String appealType = null;
-        if (applicationDTO.getCrownCourtOverviewDTO() != null
-                && applicationDTO.getCrownCourtOverviewDTO().getAppealDTO() != null
-                && applicationDTO.getCrownCourtOverviewDTO().getAppealDTO().getAppealTypeDTO() != null)
-            appealType = applicationDTO.getCrownCourtOverviewDTO().getAppealDTO().getAppealTypeDTO().getCode();
-        return appealType;
-    }
-
-    public static String getOutcome(ApplicationDTO applicationDTO) {
-        String outcome = null;
-        if ((applicationDTO.getCrownCourtOverviewDTO() != null && applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO() != null)
-                && (applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getCcOutcome() != null)) {
-                outcome = applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getCcOutcome().getOutcome();
-            }
-        return outcome;
-    }
-
-    public static String getFeeLevel(ApplicationDTO applicationDTO) {
-        String feeLevel = null;
-        if (applicationDTO.getCrownCourtOverviewDTO() != null
-                && applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO() != null
-                && applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getEvidenceProvisionFee() != null) {
-            feeLevel = applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getEvidenceProvisionFee().getFeeLevel();
-        }
-        return feeLevel;
-    }
-
-    private static boolean isEquals(Object s1, Object s2) {
-        return Objects.equals(s1, s2);
-    }
 
     public void validate(WorkflowRequest request, RepOrderDTO repOrderDTO) {
         validateApplicationTimestamp(request, repOrderDTO);
@@ -139,98 +93,6 @@ public class ValidationService {
         }
 
         return true;
-    }
-
-    public void updateSendToCCLF(WorkflowRequest request, RepOrderDTO repOrderDTO, String action) {
-        ApplicationDTO applicationDto = request.getApplicationDTO();
-        if (applicationDto == null || applicationDto.getApplicantDTO() == null || repOrderDTO == null || repOrderDTO.getId() == null) {
-            throw new ValidationException("Valid ApplicationDTO and RepOrderDTO is required");
-        }
-
-        String lAction = (action != null) ? action : "UPDATE";
-        Date cclfDate = parseDate("2021-04-01");
-        Date decisionDate = applicationDto.getDecisionDate();
-
-        if ((decisionDate.after(cclfDate) || decisionDate.equals(cclfDate)) &&
-                (lAction.equals("CREATE") || !compareRepOrderAndApplicationDTO(repOrderDTO, applicationDto))) {
-            SendToCCLFDTO sendToCCLFDTO = SendToCCLFDTO.builder().repId(repOrderDTO.getId())
-                    .applId(applicationDto.getApplicantDTO().getId())
-                    .applHistoryId(applicationDto.getApplicantDTO().getApplicantHistoryId()).build();
-            maatCourtDataApiService.updateSendToCCLF(sendToCCLFDTO);
-        }
-    }
-
-    public boolean compareRepOrderAndApplicationDTO(RepOrderDTO repOrderDTO, ApplicationDTO applicationDTO) {
-        String acctNumber = applicationDTO.getSupplierDTO() != null ? applicationDTO.getSupplierDTO().getAccountNumber() : null;
-        String feeLevel = getFeeLevel(applicationDTO);
-        LocalDate ccRepOrderDate = getRepOrderDate(applicationDTO);
-        LocalDate ccWithdrawalDate = getWithDrawalDate(applicationDTO);
-        String outcome = getOutcome(applicationDTO);
-        String court = applicationDTO.getMagsCourtDTO() != null ? applicationDTO.getMagsCourtDTO().getCourt() : null;
-        String magsOutcome = applicationDTO.getMagsOutcomeDTO() != null ? applicationDTO.getMagsOutcomeDTO().getOutcome() : null;
-        String offenceType = applicationDTO.getOffenceDTO() != null ? applicationDTO.getOffenceDTO().getOffenceType() : null;
-        Long applicantHistoryId = (applicationDTO.getApplicantDTO() != null
-                && applicationDTO.getApplicantDTO().getApplicantHistoryId() != null) ? applicationDTO.getApplicantDTO().getApplicantHistoryId() : -1;
-        String status = applicationDTO.getStatusDTO() != null ? applicationDTO.getStatusDTO().getStatus() : null;
-        String appealType = getAppealType(applicationDTO);
-        LocalDate lDateReceived = convertToLocalDateViaSqlDate(applicationDTO.getDateReceived());
-        LocalDate lCommittalDate = convertToLocalDateViaSqlDate(applicationDTO.getCommittalDate());
-
-        Long rApplicantHistoryId = repOrderDTO.getApplicantHistoryId() != null ? Long.valueOf(repOrderDTO.getApplicantHistoryId()) : -1;
-        String rOutcome = getRepOrderCcOutcome(repOrderDTO);
-
-        return isEquals(repOrderDTO.getArrestSummonsNo(), applicationDTO.getArrestSummonsNo())
-                && isEquals(repOrderDTO.getSuppAccountCode(), acctNumber)
-                && isEquals(repOrderDTO.getEvidenceFeeLevel(), feeLevel)
-                && isEquals(repOrderDTO.getMacoCourt(), court)
-                && isEquals(repOrderDTO.getMagsOutcome(), magsOutcome)
-                && isEquals(repOrderDTO.getDateReceived(), lDateReceived)
-                && isEquals(repOrderDTO.getCrownRepOrderDate(), ccRepOrderDate)
-                && isEquals(repOrderDTO.getOftyOffenceType(), offenceType)
-                && isEquals(repOrderDTO.getCrownWithdrawalDate(), ccWithdrawalDate)
-                && isEquals(repOrderDTO.getCaseId(), applicationDTO.getCaseId())
-                && isEquals(repOrderDTO.getCommittalDate(), lCommittalDate)
-                && isEquals(rApplicantHistoryId, applicantHistoryId)
-                && isEquals(repOrderDTO.getRorsStatus(), status)
-                && isEquals(rOutcome, outcome)
-                && isEquals(repOrderDTO.getAppealTypeCode(), appealType);
-    }
-
-    public LocalDate getWithDrawalDate(ApplicationDTO applicationDTO) {
-        LocalDate ccWithDrawalDate = null;
-        if ((applicationDTO.getCrownCourtOverviewDTO() != null
-                && applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO() != null)
-            && (applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getCcWithDrawalDate() != null)) {
-                ccWithDrawalDate = convertToLocalDateViaSqlDate(applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getCcWithDrawalDate());
-            }
-        return ccWithDrawalDate;
-    }
-
-    public LocalDate getRepOrderDate(ApplicationDTO applicationDTO) {
-        LocalDate ccRepOrderDate = null;
-        if ((applicationDTO.getCrownCourtOverviewDTO() != null
-                && applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO() != null)
-            && (applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getCcRepOrderDate() != null)) {
-                ccRepOrderDate = convertToLocalDateViaSqlDate(applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getCcRepOrderDate());
-            }
-        return ccRepOrderDate;
-    }
-
-    public Date parseDate(String dateString) {
-        if (dateString == null) return null;
-        try {
-            return new SimpleDateFormat("yyyy-MM-dd").parse(dateString);
-        } catch (ParseException e) {
-            throw new APIClientException("Invalid date format");
-        }
-    }
-
-
-    public LocalDate convertToLocalDateViaSqlDate(Date dateToConvert) {
-        if (dateToConvert == null) return null;
-        return dateToConvert.toInstant()
-                .atZone(ZoneId.systemDefault())
-                .toLocalDate();
     }
