+++ maat-orchestration/src/test/java/uk/gov/justice/laa/crime/orchestration/service/ValidationServiceTest.java
@@ -9,51 +9,33 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.justice.laa.crime.commons.exception.APIClientException;
-import uk.gov.justice.laa.crime.enums.AppealType;
-import uk.gov.justice.laa.crime.enums.CrownCourtOutcome;
 import uk.gov.justice.laa.crime.enums.NewWorkReason;
 import uk.gov.justice.laa.crime.enums.RepOrderStatus;
 import uk.gov.justice.laa.crime.exception.ValidationException;
 import uk.gov.justice.laa.crime.orchestration.data.builder.TestModelDataBuilder;
 import uk.gov.justice.laa.crime.orchestration.dto.WorkflowRequest;
-import uk.gov.justice.laa.crime.orchestration.dto.maat.ApplicationDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.maat_api.RepOrderDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.validation.UserSummaryDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.validation.UserValidationDTO;
 import uk.gov.justice.laa.crime.orchestration.enums.Action;
 import uk.gov.justice.laa.crime.orchestration.exception.CrimeValidationException;
-import uk.gov.justice.laa.crime.orchestration.service.api.MaatCourtDataApiService;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.util.Date;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.justice.laa.crime.orchestration.data.builder.TestModelDataBuilder.*;
+import static uk.gov.justice.laa.crime.orchestration.data.builder.TestModelDataBuilder.TEST_USER_SESSION;
 import static uk.gov.justice.laa.crime.orchestration.service.ValidationService.*;
 
 @ExtendWith(MockitoExtension.class)
 class ValidationServiceTest {
-    private static final LocalDateTime CC_WITHDRAWAL_DATETIME = LocalDateTime.of(2022, 10, 14, 0, 0, 0);
-    private static final LocalDateTime CC_REP_ORDER_DATETIME = LocalDateTime.of(2022, 10, 13, 0, 0, 0);
-
     @InjectMocks
     private ValidationService validationService;
     @Mock
     private MaatCourtDataService maatCourtDataService;
-
-    @Mock
-    private MaatCourtDataApiService maatCourtDataApiService;
 
     private static Stream<Arguments> validateApplicationTimestamp() {
         return Stream.of(
@@ -338,201 +320,5 @@ class ValidationServiceTest {
         Boolean isUserActionValid =
                 validationService.isUserActionValid(userValidationDTO);
         assertTrue(isUserActionValid);
-    }
-
-    @Test
-    void givenInValidInputWithoutApplicationDTO_whenUpdateSendToCCLFIsInvoked_thenValidationExceptionIsThrown() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        workflowRequest.setApplicationDTO(null);
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        String action = "action";
-        assertThatThrownBy(() -> validationService.updateSendToCCLF(workflowRequest, repOrderDTO, action))
-                .isInstanceOf(ValidationException.class)
-                .hasMessage("Valid ApplicationDTO and RepOrderDTO is required");
-    }
-
-    @Test
-    void givenInValidInputWithoutApplicantDTO_whenUpdateSendToCCLFIsInvoked_thenValidationExceptionIsThrown() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        workflowRequest.getApplicationDTO().setApplicantDTO(null);
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        String action = "action";
-        assertThatThrownBy(() -> validationService.updateSendToCCLF(workflowRequest, repOrderDTO, action))
-                .isInstanceOf(ValidationException.class)
-                .hasMessage("Valid ApplicationDTO and RepOrderDTO is required");
-    }
-
-    @Test
-    void givenInValidInputWithoutRepId_whenUpdateSendToCCLFIsInvoked_thenValidationExceptionIsThrown() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        repOrderDTO.setId(null);
-        String action = "action";
-        assertThatThrownBy(() -> validationService.updateSendToCCLF(workflowRequest, repOrderDTO, action))
-                .isInstanceOf(ValidationException.class)
-                .hasMessage("Valid ApplicationDTO and RepOrderDTO is required");
-    }
-
-    @Test
-    void givenInValidInputWithoutRepOrderDTO_whenUpdateSendToCCLFIsInvoked_thenValidationExceptionIsThrown() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        String action = "action";
-        assertThatThrownBy(() -> validationService.updateSendToCCLF(workflowRequest, null, action))
-                .isInstanceOf(ValidationException.class)
-                .hasMessage("Valid ApplicationDTO and RepOrderDTO is required");
-    }
-
-
-    @Test
-    void givenValidInputWithUpdateAction_whenUpdateSendToCCLFIsInvoked_thenNoExceptionIsThrown() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        String action = "UPDATE";
-        assertDoesNotThrow(() -> validationService.updateSendToCCLF(workflowRequest, repOrderDTO, action));
-    }
-
-    @Test
-    void givenValidInputWithNullAction_whenUpdateSendToCCLFIsInvoked_thenNoExceptionIsThrown() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        assertDoesNotThrow(() -> validationService.updateSendToCCLF(workflowRequest, repOrderDTO, null));
-    }
-
-    @Test
-    void givenValidInputWithBeforeDecisionDate_whenUpdateSendToCCLFIsInvoked_thenNoExceptionIsThrown() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        LocalDateTime DECISION_DATETIME = LocalDateTime.of(2000, 10, 13, 0, 0, 0);
-        workflowRequest.getApplicationDTO().setDecisionDate(Date.from(DECISION_DATETIME.atZone(ZoneId.systemDefault()).toInstant()));
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        assertDoesNotThrow(() -> validationService.updateSendToCCLF(workflowRequest, repOrderDTO, null));
-    }
-
-    @Test
-    void givenValidInputWithDifferentObject_whenCompareRepOrderAndApplicationDTOIsInvoked_thenOKResponseIsReturned() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        assertThat(validationService.compareRepOrderAndApplicationDTO(repOrderDTO, workflowRequest.getApplicationDTO())).isFalse();
-    }
-
-    @Test
-    void givenValidInputWithSameAttributes_whenCompareRepOrderAndApplicationDTOIsInvoked_thenOKResponseIsReturned() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        ApplicationDTO applicationDTO = getTestApplicationDTO(workflowRequest);
-        RepOrderDTO repOrderDTO = getTestRepOrderDTO(applicationDTO);
-        assertThat(validationService.compareRepOrderAndApplicationDTO(repOrderDTO, workflowRequest.getApplicationDTO())).isTrue();
-    }
-
-
-    @Test
-    void givenValidInput_whenUpdateSendToCCLFIsInvoked_thenOKResponseIsReturned() {
-        WorkflowRequest workflowRequest = TestModelDataBuilder.buildWorkFlowRequest();
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        String action = "CREATE";
-        when(maatCourtDataApiService.updateSendToCCLF(any())).thenReturn(repOrderDTO);
-        assertDoesNotThrow(() -> validationService.updateSendToCCLF(workflowRequest, repOrderDTO, action));
-        verify(maatCourtDataApiService).updateSendToCCLF(any());
-    }
-
-    @Test
-    void givenValidInput_whenGetWithdrawalDateIsInvoked_thenOKResponseIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().setCcWithDrawalDate(Date.from(CC_WITHDRAWAL_DATETIME.atZone(ZoneId.systemDefault()).toInstant()));
-        assertThat(validationService.getWithDrawalDate(applicationDTO)).isEqualTo(LocalDate.from(CC_WITHDRAWAL_DATETIME));
-    }
-
-    @Test
-    void givenInputWithOutCCO_whenGetWithdrawalDateIsInvoked_thenNullIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.setCrownCourtOverviewDTO(null);
-        assertNull(validationService.getWithDrawalDate(applicationDTO));
-    }
-
-    @Test
-    void givenInputWithOutCCSummary_whenGetWithdrawalDateIsInvoked_thenNullIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().setCrownCourtSummaryDTO(null);
-        assertNull(validationService.getWithDrawalDate(applicationDTO));
-    }
-
-    @Test
-    void givenInputWithOutCCWithdrawal_whenGetWithdrawalDateIsInvoked_thenNullIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().setCcWithDrawalDate(null);
-        assertNull(validationService.getWithDrawalDate(applicationDTO));
-    }
-
-    @Test
-    void givenValidInput_whenGetRepOrderDateIsInvoked_thenOKResponseIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().setCcRepOrderDate(Date.from(CC_REP_ORDER_DATETIME.atZone(ZoneId.systemDefault()).toInstant()));
-        assertThat(validationService.getRepOrderDate(applicationDTO)).isEqualTo(LocalDate.from(CC_REP_ORDER_DATETIME));
-    }
-
-    @Test
-    void givenInputWithOutCCO_whenGetRepOrderDateIsInvoked_thenNullIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.setCrownCourtOverviewDTO(null);
-        assertNull(validationService.getRepOrderDate(applicationDTO));
-    }
-
-    @Test
-    void givenInputWithOutCCSummary_whenGetRepOrderDateIsInvoked_thenNullIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().setCrownCourtSummaryDTO(null);
-        assertNull(validationService.getRepOrderDate(applicationDTO));
-    }
-
-    @Test
-    void givenInputWithOutCCRepOrder_whenGetRepOrderDateIsInvoked_thenNullIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().setCcRepOrderDate(null);
-        assertNull(validationService.getRepOrderDate(applicationDTO));
-    }
-
-    @Test
-    void parseValidDate() throws ParseException {
-        String date = "2023-01-01";
-        Date result = validationService.parseDate(date);
-        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-        Date expected = format.parse(date);
-        assertEquals(expected, result);
-    }
-
-    @Test()
-    void parseInvalidFormat() {
-        assertThatThrownBy(() -> validationService.parseDate("invalid")).isInstanceOf(APIClientException.class);
-    }
-
-    @Test()
-    void parseNullDate() {
-        assertNull(validationService.parseDate(null));
-    }
-
-    @Test
-    void givenValidInput_whenGetRepOrderCcOutcomeIsInvoked_thenValidResponseIsReturned() {
-        RepOrderDTO repOrderDTO = TestModelDataBuilder.buildRepOrderDTO("CURR");
-        repOrderDTO.getRepOrderCCOutcome().add(TestModelDataBuilder.getRepOrderCCOutcomeDTO());
-        assertEquals("CONVICTED", ValidationService.getRepOrderCcOutcome(repOrderDTO));
-    }
-
-    @Test
-    void givenValidInput_whenGetAppealTypeIsInvoked_thenValidResponseIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().getAppealDTO().getAppealTypeDTO().setCode(AppealType.ACS.getCode());
-        assertEquals(AppealType.ACS.getCode(), ValidationService.getAppealType(applicationDTO));
-    }
-
-    @Test
-    void givenValidInput_whenGetOutcomeIsInvoked_thenValidResponseIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().setCcOutcome(TestModelDataBuilder.getOutcomeDTO());
-        assertEquals(CrownCourtOutcome.SUCCESSFUL.toString(), ValidationService.getOutcome(applicationDTO));
-    }
-
-    @Test
-    void givenValidInput_whenGetFeeLevelIsInvoked_thenValidResponseIsReturned() {
-        ApplicationDTO applicationDTO = TestModelDataBuilder.buildWorkFlowRequest().getApplicationDTO();
-        applicationDTO.getCrownCourtOverviewDTO().getCrownCourtSummaryDTO().getEvidenceProvisionFee().setFeeLevel("TEST");
-        assertEquals("TEST", ValidationService.getFeeLevel(applicationDTO));
     }
 }
